pyproject.toml
[tool.poetry]
name = "mlops-on-gcp"
version = "0.1.0"
description = ""
authors = ["Simon Löw"]
readme = "README.md"
packages = [{include = "part1"}]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-aiplatform = "^1.20.0"
google-cloud-pipeline-components = "^1.0.32"
kfp = "^1.8.18"
scikit-learn = "^1.2.0"
pandas = "^1.5.2"
wget = "^3.2"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}
ipykernel = "^6.19.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
